Using custom workflow actions for test and deploy.
The workflow actions at vguerra/fastai-workflow actions
allow to pass the `use-requirements-file` parameter that will
use the `requirements.txt` file during the instalation of the package.

We need this because we want to be able to install pre-release
packages for `torch` and `torchvision` found in the nightly wheel
repo for pytorch.

Ref for this:
https://caremad.io/posts/2013/07/setup-vs-requirement/